add alerts

+++ client/src/pages/home/Home.tsx
@@ -1,11 +1,12 @@
 import { useEffect, useState } from "react";
 import { ComplexNavbar } from "../../components/navbar/Navbar";
 import { Task } from "../../interfaces";
-import { deleteSelectedTasks, deleteTask, getAllTasks, toggleTaskCompleted } from "../../services/task.services";
+import { deleteAllTasks, deleteSelectedTasks, deleteTask, getAllTasks, toggleTaskCompleted } from "../../services/task.services";
 import { Checkbox } from "@material-tailwind/react";
 import { PlusIcon, TrashIcon } from "@heroicons/react/24/outline";
 import { ModalCreateTask } from "./ModalCreateTask";
 import { getDays } from "../../utils";
+import Swal from "sweetalert2";
 
 const Home = () => {
     const [tasks, setTasks] = useState<Task[]>([]);
@@ -37,17 +38,75 @@ const Home = () => {
     };
 
     const handleDeleteTask = (id: string) => {
-        deleteTask(id).then(() => {
-            setTasks(tasks.filter((task) => task.task_id !== id));
-
-        });
+        Swal.fire({
+            title: 'Are you sure?',
+            text: "You will not be able to recover this task!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#10B981',
+            cancelButtonColor: '#EF4444',
+            confirmButtonText: 'Yes, delete it!'
+        }).then((result) => {
+            if (result.isConfirmed) {
+                deleteTask(id).then(() => {
+                    setTasks(tasks.filter((task) => task.task_id !== id));
+                    setSelectedTask(selectedTask.filter((taskId) => taskId !== id));
+                });
+                Swal.fire(
+                    'Deleted!',
+                    'Your task has been deleted.',
+                    'success'
+                )
+            }
+        })
     };
     const handleDeleteSelectedTasks = () => {
-        deleteSelectedTasks(selectedTask).then(() => {
-            setTasks(tasks.filter((task) => !selectedTask.includes(task.task_id)));
-            setSelectedTask([]);
-        });
+        Swal.fire({
+            title: 'Are you sure?',
+            text: "You will not be able to recover this tasks!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#10B981',
+            cancelButtonColor: '#EF4444',
+            confirmButtonText: 'Yes, delete them!'
+        }).then((result) => {
+            if (result.isConfirmed) {
+                deleteSelectedTasks(selectedTask).then(() => {
+                    setTasks(tasks.filter((task) => !selectedTask.includes(task.task_id)));
+                    setSelectedTask([]);
+                });
+                Swal.fire(
+                    'Deleted!',
+                    'Your tasks has been deleted.',
+                    'success'
+                )
+            }
+        }
+        )
     };
+    const handleDeleteAllTasks = () => {
+        Swal.fire({
+            title: 'Are you sure?',
+            text: "You will not be able to recover this tasks!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#10B981',
+            cancelButtonColor: '#EF4444',
+            confirmButtonText: 'Yes, delete them!'
+        }).then((result) => {
+            if (result.isConfirmed) {
+                deleteAllTasks().then(() => {
+                    setTasks([]);
+                    setSelectedTask([]);
+                });
+                Swal.fire(
+                    'Deleted!',
+                    'Your tasks has been deleted.',
+                    'success'
+                )
+            }
+        })
+    }
     const [open, setOpen] = useState(false);
     const handleOpen = () => setOpen(!open);
 
@@ -56,20 +115,24 @@ const Home = () => {
             <ModalCreateTask setTasks={setTasks} open={open} handler={handleOpen} />
             <ComplexNavbar />
             <main className="flex justify-center mt-10 min-h-[calc(100vh-200px)]">
-                <div className="w-full flex flex-col xl:w-3/4 xl:flex-row ">
+                <div className="w-full flex flex-col xl:w-3/4  2xl:flex-row  gap-2 ">
                     <section className="w-full" >
-                        <div className="flex justify-between w-full border-b p-4 ">
-                            <div className="flex flex-col justify-center ">
+                        <div className="flex flex-col lg:flex-row justify-between w-full border-b p-4 ">
+                            <div className="flex flex-col  justify-center mb-4 xl:mb-0 ">
                                 <h1 className="text-3xl font-bold">To Do</h1>
                                 <p className="text-sm">Tasks to do</p>
                             </div>
-                            <div className="flex items-center">
+                            <div className="flex gap-1  items-center ">
                                 {/* add task*/}
-                                <button onClick={handleOpen} className="bg-gray-500 hover:bg-green-400 text-white px-4 py-2 rounded-md mr-2 flex items-center gap-2">
+                                <button onClick={handleOpen} className="bg-gray-500 text-sm hover:bg-green-400 text-white p-2 rounded-md mr-2 flex flex-grow lg:flex-grow-0 items-center gap-2">
                                     <PlusIcon className="h-6 w-6" />
                                     Add task
                                 </button>
-                                {selectedTask.length > 0 && <button onClick={handleDeleteSelectedTasks} className="bg-gray-500 hover:bg-red-400 text-white px-4 py-2 rounded-md mr-2 flex items-center gap-2">
+                                <button onClick={handleDeleteAllTasks} className="bg-gray-500 text-sm hover:bg-red-400 text-white flex-grow p-2 lg:flex-grow-0  rounded-md mr-2 flex  items-center gap-2">
+                                    <TrashIcon className="h-6 w-6" />
+                                    Clear board
+                                </button>
+                                {selectedTask.length > 0 && <button onClick={handleDeleteSelectedTasks} className="bg-gray-500 text-sm hover:bg-red-400  flex-grow lg:flex-grow-0 text-white p-2 rounded-md mr-2 flex items-center gap-2">
                                     <TrashIcon className="h-6 w-6" />
                                     {`Delete ${selectedTask.length} tasks`}
                                 </button>}
@@ -122,7 +185,7 @@ const Home = () => {
                             {tasks.filter(task => task.completed).length > 0 ? tasks.filter((task) => task.completed).map((task) => {
                                 return (
                                     <li
-                                        className={`flex items-center w-full shadow-md py-4 px-2 mb-4 cursor-pointer ${isTaskSelected(task.task_id) ? "bg-blue-gray-50" : "bg-gray-100"} opacity-80 rounded-md `} key={task.task_id}>
+                                        className={`flex items-center w-full shadow-md py-4 px-2 mb-4  ${isTaskSelected(task.task_id) ? "bg-blue-gray-50" : "bg-gray-100"} opacity-80 rounded-md `} key={task.task_id}>
                                         <div className="border-r border-gray-300 p-2">
 
                                             <Checkbox
@@ -140,9 +203,9 @@ const Home = () => {
                                             <small className="text-xs text-gray-600 ">Description</small>
                                             <p className="text-sm">{task.description}</p>
                                         </div>
-                                        <div className="p-4 border-l">
+                                        <button onClick={() => handleDeleteTask(task.task_id)} className="p-4 border-l">
                                             <TrashIcon className="h-6 w-6 text-red-500 cursor-pointer" />
-                                        </div>
+                                        </button>
                                     </li>
                                 )
                             }) : <div className="flex  items-center justify-center w-full h-full gap-2">
@@ -155,7 +218,7 @@ const Home = () => {
             {/* footer */}
             <small className="text-center block w-full mt-10 text-gray-500">
                 &copy; 2023 DoLister. All rights reserved.
-                Powered by <a href="https://github.com/fallenxl" target="_blank"  className="underline hover:text-gray-600">FallEn</a>
+                Powered by <a href="https://github.com/fallenxl" target="_blank" className="underline hover:text-gray-600">FallEn</a>
             </small>
         </div>
     );
